+++ server/src/main/java/meghanada/project/gradle/Configuration.java
@@ -1,0 +1,9 @@
+package meghanada.project.gradle;
+
+import java.util.List;
+
+public interface Configuration {
+  String getName();
+
+  List<Dependency> getDependencies();
+}