+++ server/src/main/java/meghanada/server/emacs/EmacsServer.java
@@ -31,6 +31,8 @@ import meghanada.server.formatter.SExprOutputFormatter;
 import meghanada.session.Session;
 import meghanada.session.SessionEventBus;
 import meghanada.system.Executor;
+import meghanada.telemetry.ErrorReporter;
+import meghanada.utils.FileUtils;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -349,7 +351,7 @@ public class EmacsServer implements Server {
                 })
             .getMatch();
 
-    log.info("receive {}. elapsed:{}", argList, stopwatch.stop());
+    log.info("receive {}. {}", convertArgs(argList), stopwatch.stop());
     return result;
   }
 
@@ -372,6 +374,7 @@ public class EmacsServer implements Server {
       this.accept();
     } catch (Throwable e) {
       log.catching(e);
+      ErrorReporter.report(e);
     } finally {
       try {
         this.serverSocket.close();
@@ -463,6 +466,18 @@ public class EmacsServer implements Server {
       return new SExprOutputFormatter();
     }
     throw new UnsupportedOperationException("not support format");
+  }
+
+  private List<String> convertArgs(List<String> argList) {
+    return argList.stream()
+        .map(
+            s -> {
+              if (s.endsWith(".java")) {
+                return FileUtils.shortenPath(s);
+              }
+              return s;
+            })
+        .collect(Collectors.toList());
   }
 
   private enum OUTPUT {