+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# Next Release
+
+# 1.3.1 (2020-08-15)
+
+## Added
+
+- Add the function of sending an error report.
+
+## Changed
+
+- Update gradle to 6.6.
+- Fix to use the latest library.
+- Improve compile error logs.
+- Improve detection of dependent libraries in gradle projects.
+- Fix to keep a class cache for each project.
+- Shorten the File path in the log
+
+## Fixed
+
 # 1.3.0 (2020-05-07)
 
 ## Added